pyproject.toml
# Poetry
[tool.poetry]
name = "priscillafx-website"
version = "0.1.0"
description = "Official website of Priscilla FX"
authors = ["masajinobe-ef <priscilla.effects@gmail.com>"]
license = "GPL-3.0"
readme = "README.md"
packages = [{ include = "*", from = "src/backend" }]

# Deps
[tool.poetry.dependencies]
python = "^3.10"
python-multipart = "^0.0.9"

fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
gunicorn = "^22.0.0"
jinja2 = "^3.1.3"

psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.29"
sqlmodel = "^0.0.16"
alembic = "^1.13.1"

aioredis = "^2.0.1"
fastapi-cache2 = { extras = ["redis"], version = "^0.2.1" }

fastapi-users = { extras = ["sqlalchemy"], version = "^13.0.0" }
cryptography = "^42.0.8"

celery = "^5.4.0"
flower = "^2.0.1"

loguru = "^0.7.2"

# Dev-Deps
[tool.poetry.group.dev.dependencies]
commitizen = "^3.27.0"


# Ruff
[tool.ruff]
line-length = 79
indent-width = 4

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

[tool.ruff.lint]
ignore = ["F403"]

# Commitizen
[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "poetry"
version = "0.1.0"
update_changelog_on_bump = true
major_version_zero = true


# Poetry build system
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
